Issue #1202: Improve setting IntegerObject values
- Check if value for IntegerObject is in 32-bit range
- Raise ValueError for out of range values
- Add tests for the supported range and out of range
